backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
Pillow
pydantic